Remove needless `wasm-builder` feature and merge runtime module into lib.rs in evm-runtime
