+++ domains/runtime/evm/build.rs
@@ -3,7 +3,6 @@ fn main() {
     {
         substrate_wasm_builder::WasmBuilder::new()
             .with_current_project()
-            .enable_feature("wasm-builder")
             .export_heap_base()
             .import_memory()
             .build();
